+++ frb_codegen/src/library/codegen/ir/hir/raw/mod.rs
@@ -1,7 +1,3 @@
 pub(crate) mod crates;
+pub(crate) mod pack;
 
-use crate::codegen::ir::hir::raw::crates::HirRawCrate;
-
-pub(crate) struct HirRawPack {
-    pub crates: Vec<HirRawCrate>,
-}
