+++ frb_codegen/src/library/codegen/ir/hir/raw/pack.rs
@@ -1,0 +1,5 @@
+use crate::codegen::ir::hir::raw::crates::HirRawCrate;
+
+pub(crate) struct HirRawPack {
+    pub crates: Vec<HirRawCrate>,
+}
